Refactor: export public chat mediator to a new file
