+++ back-end/src/chats/publicChatMediator.ts
@@ -1,23 +1,23 @@
 import { Server, Socket } from 'socket.io';
+import { Client } from './client';
 import { Message } from '../models/message';
 import { User } from '../models/user';
-import { Client } from './client';
 
-type ChatMessage = {
+export type ChatMessage = {
   id: string;
   userName: string;
   message: string;
 };
 
-type UserType = {
+export type UserType = {
   id: string;
   userName: string;
 };
 
-const onlineUsers: UserType[] = [];
+export const onlineUsers: UserType[] = [];
 
 // Mediator pattern: handling of chat logic with classes
-class PublicChatMediator {
+export class PublicChatMediator {
   private static instance: PublicChatMediator;
   private clients: Client[] = [];
 
@@ -97,43 +97,3 @@ class PublicChatMediator {
     this.io.emit('onlineUsers', onlineUsers);
   }
 }
-
-export const publicChatHandler = (io: Server, socket: Socket) => {
-  const chatMediator = PublicChatMediator.getInstance(io);
-  const client = new Client(socket);
-
-  chatMediator.addClient(client);
-
-  socket.on('getPreviousMessages', () => {
-    chatMediator.sendPreviousMessages(client);
-  });
-
-  socket.on('chatMessage', (msg: ChatMessage) => {
-    chatMediator.broadcastMessage(msg, socket);
-  });
-
-  socket.on('joinChat', (userName: string) => {
-    const user = { id: socket.id, userName };
-
-    const existingUserIndex = onlineUsers.findIndex(
-      (u) => u.userName === userName
-    );
-    if (existingUserIndex === -1) {
-      onlineUsers.push(user);
-    } else {
-      onlineUsers[existingUserIndex].id = socket.id;
-    }
-
-    chatMediator.broadcastOnlineUsers();
-  });
-
-  socket.on('disconnect', () => {
-    chatMediator.removeClient(socket.id);
-
-    const index = onlineUsers.findIndex((user) => user.id === socket.id);
-    if (index !== -1) {
-      onlineUsers.splice(index, 1);
-      chatMediator.broadcastOnlineUsers();
-    }
-  });
-};
